+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Change log
 - Promoted ``_reset_passwords`` to ``reset-pw``.
 - Removed the explicit activation of pip's 2020 resolver from pip
   invocations, it is the default now.
+- Added ``nine-restart`` to restart the application server.
 
 
 `1.0.20201110`_